refactor(runtime): extract network fault module from transport coordination (#3187)
